GPII-2379:  Remove previous Testem harness in favor of gpii-testem.
