+++ package.json
@@ -5,24 +5,22 @@
   "license": "BSD-3-Clause",
   "repository": "https://github.com/GPII/gpii-binder",
   "scripts": {
-    "instrument": "node node_modules/istanbul/lib/cli.js instrument --no-compact --output instrumented src",
-    "pretest": "node node_modules/rimraf/bin.js coverage instrumented && npm run instrument",
-    "test": "node node_modules/testem/testem.js ci --skip PhantomJS,Chromium,Safari",
-    "posttest": "node node_modules/istanbul/lib/cli.js report --root coverage --dir reports text-summary html"
+    "test": "node node_modules/testem/testem.js --file tests/testem.js ci --skip PhantomJS,Chromium,Safari"
   },
   "dependencies": {
     "infusion": "3.0.0-dev.20170127T130413Z.103de6e"
   },
   "devDependencies": {
-    "eslint": "3.14.1",
+    "eslint": "3.19.0",
     "eslint-config-fluid": "1.1.0",
     "fluid-grunt-eslint": "18.1.2",
+    "gpii-testem": "1.0.0-dev.20170411T125618Z.4ec8037",
     "grunt": "1.0.1",
     "grunt-jsonlint": "1.1.0",
     "istanbul": "0.4.5",
     "node-jqunit": "1.1.4",
     "qunitjs": "2.1.1",
-    "rimraf": "2.5.4",
-    "testem": "1.14.3"
+    "rimraf": "2.6.1",
+    "testem": "1.13.0"
   }
 }
