+++ tests/static/all-tests.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <title>Fluid framework complete tests runner</title>
 
-    <script type="text/javascript" src="./js/testem/coverage.js"></script>
+    <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
 
     <link rel="stylesheet" href="../../node_modules/node-jqunit/node_modules/infusion/tests/lib/qunit/css/qunit.css"/>
     <link rel="stylesheet" href="../../node_modules/node-jqunit/node_modules/infusion/tests/lib/qunit/addons/composite/qunit-composite.css"/>
