+++ tests/static/tests-binder-radio.html
@@ -29,7 +29,7 @@
         <script type="text/javascript" src="./js/tests-fixtures.js"></script>
 
         <!-- Bring in client library being tested -->
-        <script type="text/javascript" src="./js/testem/coverage.js"></script>
+        <script type="text/javascript" src="/coverage/client/coverageSender.js"></script>
         <script type="text/javascript" src="../../src/js/binder.js"></script>
 
         <link rel="stylesheet" href="../../node_modules/node-jqunit/node_modules/infusion/tests/lib/qunit/css/qunit.css"/>
